updated structure commit
