+++ frontend/src/routes/AppRoutes.js
@@ -1,6 +1,7 @@
 import React, { lazy, Suspense } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
+const Auth = lazy(() => import("../pages/auth"))
 const Home = lazy(() => import("../pages/Home"));
 const AddItems = lazy(() => import("../pages/AddItems"));
 const Stocks = lazy(() => import("../pages/Stocks"));
@@ -11,7 +12,8 @@ const AppRoutes = () => {
     <Router>
       <Suspense fallback={<div>Loading...</div>}>
         <Routes>
-          <Route path="/" element={<Home />} />
+          <Route path="/" element={<Auth />} />
+          <Route path="/home" element={<Home />} />
           <Route path="/addItems" element={<AddItems />} />
           <Route path="/stocks" element={<Stocks />} />
           <Route path="*" element={<NotFound />} /> {/* 404 Page */}
